image/requirements.txt
Django==3.2.19
djangorestframework==3.14.0
requests==2.31.0
python-dotenv==1.0.1
django-cors-headers==4.3.1
